[move] ConfiguredActorSystemProviderをseichiassist以下に移動する
